services/scheduler-service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eve-scheduler-service"
version = "1.0.0"
description = "EVE Co-Pilot Scheduler Service - Centralized Job Scheduling"
requires-python = ">=3.11"
dependencies = [
    "eve-shared",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "apscheduler>=3.10.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.0",
    "httpx>=0.25.0",
    # Legacy job dependencies
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.3.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
